requirements.txt
httpx[http2]
beautifulsoup4
lxml
click
rich
jinja2
aiofiles
urllib3
structlog
pyyaml
fastapi
uvicorn[standard]
sqlalchemy
pydantic
websockets
python-multipart